% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso.bag.R, R/lassoBagAddGPD.R
\name{Lasso.bag}
\alias{Lasso.bag}
\title{LASSO-bagging: a lasso based variable selecting CART framework}
\usage{
Lasso.bag(mat, out.mat, bootN = 1000, imputeN = 1000,
  imputeN.max = 2000, permut.increase = 100, boot.rep = TRUE,
  a.family = c("gaussian", "binomial", "poisson", "multinomial", "cox",
  "mgaussian"), parallel = F)

Lasso.bag(mat, out.mat, bootN = 1000, imputeN = 1000,
  imputeN.max = 2000, permut.increase = 100, boot.rep = TRUE,
  a.family = c("gaussian", "binomial", "poisson", "multinomial", "cox",
  "mgaussian"), parallel = F)
}
\arguments{
\item{mat}{sample matrix that each column represent a variable and rows represent sample data points}

\item{out.mat}{vector with the same length as the sample size from `mat`}

\item{a.family}{a string vector}

\item{universe}{total number of all strings that vec1 and vec2 comes from}

\item{mat}{sample matrix that each column represent a variable and rows represent sample data points}

\item{out.mat}{vector with the same length as the sample size from `mat`}

\item{a.family}{a string vector}

\item{universe}{total number of all strings that vec1 and vec2 comes from}
}
\value{
a P value

a P value
}
\description{
LASSO-bagging: a lasso based variable selecting CART framework

LASSO-bagging: a lasso based variable selecting CART framework
}
