% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoBagAddGPD.R
\docType{package}
\name{LassoBag}
\alias{LassoBag}
\alias{Lasso.bag}
\title{LassoBag: a lasso based variable selecting CART framework}
\usage{
Lasso.bag(
  mat,
  out.mat,
  base.exist = F,
  bootN = 1000,
  imputeN = 1000,
  imputeN.max = 2000,
  permut.increase = 100,
  boot.rep = TRUE,
  a.family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  additional.info = NULL,
  bagFreq.sigMethod = "simple estimation",
  local.min.BIC = TRUE,
  use.gpd = F,
  fit.pareto = "gd",
  parallel = F,
  n.cores = 1,
  rd.seed = 89757,
  nfolds = 4,
  lambda.type = "lambda.1se",
  plot.freq = "full",
  plot.out = F,
  do.plot = F,
  output_dir = NA,
  filter_method = "auto",
  filter_inbag = FALSE,
  filter_thres_method = "traditional fdr",
  filter_thres_fdr = 0.05,
  filter_rank_cutoff = 0.05,
  FilterInitPermutT = 100,
  FilterIncPermutSt = 100,
  FilterMAXPermutT = 2000,
  filter_result_report = T,
  report_all = T,
  post.regression = TRUE,
  post.LASSO = TRUE,
  pvalue.cutoff = 0.05
)
}
\arguments{
\item{mat}{sample matrix that each column represent a variable and rows represent sample data points, all the entries in it should be numeric.}

\item{out.mat}{vector or dataframe with two columns with the same length as the sample size from `mat`}

\item{base.exist}{indicates whether we have an existed frequency data frame. We set it to be FALSE under most condition. If you don't know what it means, leave it to be FALSE(default)}

\item{bootN}{the size of resampled sample, only valid when permutation set to TRUE}

\item{imputeN}{the initial permutation times, only valid when permutation set to TRUE}

\item{imputeN.max}{the max permutation times. Regardless of whether p has meet the requirement,, only valid when permutation set to TRUE}

\item{permut.increase}{if the initial imputeN times of permutation doesn't meet the requirement, then we add ‘permut.increase times of permutation??? to get more random/permutation values, only valid when permutation set to TRUE}

\item{boot.rep}{whether :"sampling with return" or not, only valid when permutation set to TRUE}

\item{a.family}{a string vector to determine the data type of out.mat}

\item{fit.pareto}{the method of fitting Generalized Pareto Distribution, alternative choice is "gd", for gradient descend}

\item{parallel}{whether the script run in parallel, you need to set n.cores in case this package conquers all your cpu resource}

\item{n.cores}{how many cores/process to be assigned for this function, in Windows, you have to set this to 1}

\item{rd.seed}{it is the random seed of this function, in case some of the experiments need to be reappeared}

\item{plot.freq}{whether to show all the non-zero frequency in the final barplot or not. If "full", all the features(including zero frequency) will be plotted. If "part", all the non-zero features will be plotted. If "not", will not print the plot.}

\item{plot.out}{the path or file's name to save the plot. If set to FALSE, no plot will be output. If you run this function in Linux command line, you don't have to set this param for the plot.freq will output your plot to your current working directory with name "Rplot.pdf".Default to FALSE.}

\item{do.plot}{whether to print a barplot to show the frequency of the output. This package will show you a barplot when it is TRUE. If it is FALSE, the two plot.* params above will be invalid then.}

\item{permutation}{to decide whether to do permutation test, if set FALSE, no p value returns}
}
\value{
a dataframe that contains the frequency, the p value and the adjusted p value of each feature(if you set permutation=T)
}
\examples{
NULL
}
