% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LessPermutation.R
\name{LessPermutation}
\alias{LessPermutation}
\title{LessPermutation calculating precised p value with less permutation by fitting a General Pareto Distribution(GPD)}
\usage{
LessPermutation(X, x0, fitting.method = "mle", search.step = 0.01,
  fit.cutoff = 0.05, when.to.fit = 0.05)
}
\description{
LessPermutation calculating precised p value with less permutation by fitting a General Pareto Distribution(GPD)
}
