% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoBagAddGPD.R
\name{Lasso.bag}
\alias{Lasso.bag}
\title{Lasso.bag: one-step main function of LassoBag framework}
\usage{
Lasso.bag(
  mat,
  out.mat,
  observed.fre = NULL,
  bootN = 1000,
  boot.rep = TRUE,
  a.family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  additional.covariate = NULL,
  bagFreq.sigMethod = "PST",
  kneedle.S = 10,
  auto.loose = TRUE,
  loosing.factor = 0.5,
  min.S = 0.1,
  use.gpd = FALSE,
  fit.pareto = "gd",
  imputeN = 1000,
  imputeN.max = 2000,
  permut.increase = 100,
  parallel = FALSE,
  n.cores = 1,
  rd.seed = 10867,
  nfolds = 4,
  lambda.type = "lambda.1se",
  plot.freq = "part",
  plot.out = FALSE,
  do.plot = TRUE,
  output.dir = NA,
  filter.method = "auto",
  inbag.filter = TRUE,
  filter.thres.method = "fdr",
  filter.thres.P = 0.05,
  filter.rank.cutoff = 0.05,
  filter.min.features = -Inf,
  filter.max.features = Inf,
  filter.result.report = TRUE,
  filter.report.all.variates = TRUE,
  post.regression = FALSE,
  post.LASSO = FALSE,
  pvalue.cutoff = 0.05,
  used.elbow.point = "middle"
)
}
\arguments{
\item{mat}{sample matrix that each column represent a variable and rows represent sample data points, all the entries in it should be numeric.}

\item{out.mat}{vector or dataframe with the same rows as the sample size of `mat`.}

\item{observed.fre}{dataframe with columns 'variate' and 'Frequency', which can be obtained from existed LASSOBag results for re-analysis. A warning will be issued if the variates in `observed.fre` not found in `mat`, and these variates will be excluded.}

\item{bootN}{the size of re-sampled samples for bagging, default 1000; smaller consumes less processing time but may not get robust results.}

\item{boot.rep}{whether sampling with return or not in the bagging procedure}

\item{a.family}{a character determine the data type of out.mat, the same used in \code{\link[glmnet]{glmnet}}.}

\item{additional.covariate}{provide additional covariate(s) to build the cox model, only valid in Cox method (`a.family` == "cox"); a data.frame with same rows as `mat`}

\item{bagFreq.sigMethod}{a character to determine the cut-off point determination method for the importance measure (i.e. the observed selected frequency). Supported methods are "Parametrical Statistical Test" (abbr. "PST"), "Curve Elbow Point Detection" ("CEP") and "Permutation Test" ("PERT"). The default and preferable method is "CEP".}

\item{kneedle.S}{numeric, an important parameter that determines how aggressive the elbow points on the curve to be called, smaller means more aggressive and may find more elbow points. Default `kneedle.S`=5 seems fine, but feel free to try other values. The selection of `kneedle.S` should be based on the shape of observed frequency curve. It is suggested to use larger S first.}

\item{auto.loose}{if TRUE, will reduce `kneedle.S` in case no elbow point is found with the set `kneedle.S`; only valid when `bagFreq.sigMethod` is "Curve Elbow Point Detection" ("CEP").}

\item{loosing.factor}{a numeric value range in (0,1), which `kneedle.S` is multiplied by to reduce itself; only valid when `auto.loose` set to TRUE.}

\item{min.S}{a numeric value determines the minimal value that `kneedle.S` will be loosed to; only valid when `auto.loose` set to TRUE.}

\item{use.gpd}{whether to fit Generalized Pareto Distribution to the permutation results to accelerate the process. Only valid when `bagFreq.sigMethod` is "Permutation Test" ("PERT").}

\item{fit.pareto}{the method of fitting Generalized Pareto Distribution, alternative choice is "gd", for gradient descend (only valid in "PERT" mode).}

\item{imputeN}{the initial permutation times (only valid in "PERT" mode).}

\item{imputeN.max}{the max permutation times. Regardless of whether p-value has meet the requirement (only valid in "PERT" mode).}

\item{permut.increase}{if the initial imputeN times of permutation doesn't meet the requirement, then we add ‘permut.increase times of permutation to get more random/permutation values (only valid in "PERT" mode).}

\item{parallel}{whether the script run in parallel mode; you also need to set n.cores to determine how much CPU resource to use.}

\item{n.cores}{how many cores/process to be assigned for this function; more cores used results in more resource of CPU and memory used.}

\item{rd.seed}{the random seed of this function, in case some of the experiments need to be reproduced.}

\item{nfolds}{integer > 2, how many folds to be created for n-folds cross-validation LASSO in \code{\link[glmnet]{cv.glmnet}}.}

\item{lambda.type}{character, which model should be used to obtain the variates selected in one bagging. Default is "lambda.1se" for less variates selected and lower probability being over-fitting. See the help of `cv.glmnet` for more details.}

\item{plot.freq}{whether to show all the non-zero frequency in the final barplot or not. If "full", all the features(including zero frequency) will be plotted. If "part", all the non-zero features will be plotted. If "not", will not print the plot.}

\item{plot.out}{the file's name of the frequency plot. If set to FALSE, no plot will be output. If you run this function in Linux command line, you don't have to set this param for the plot.freq will output your plot to your current working directory with name "Rplot.pdf".Default to FALSE.}

\item{do.plot}{if TRUE generate result plots.}

\item{output.dir}{the path to save result files generated by \code{\link[LassoBag]{Lasso.bag}} (if not existed, will be created). Default is NA, will save in the same space as the current working dir.}

\item{filter.method}{the filter method applied to input matrix; default is `auto`, automatically select the filter method according to the data type of `out.mat`. Specific supported methods are "pearson", "spearman", "kendall" from \code{\link{cor.test}} method, and "cox" from \code{\link{coxph}} method, and "none" (no filter applied).}

\item{inbag.filter}{if TRUE, apply filters to the re-sampled bagging samples rather than the original samples; default is TRUE.}

\item{filter.thres.method}{the method determines the threshold of importance in filters. Supported methods are "fdr" and "rank".}

\item{filter.thres.P}{if `filter.thres.method` is "fdr", use `filter.thres.P` as the (adjusted) cut-off p-value. Default is 0.05.}

\item{filter.rank.cutoff}{if `filter.thres.method` is "rank", use `filter.rank.cutoff` as the cut-off rank. Default is 0.05.}

\item{filter.min.features}{minimum important features selected by filters. Useful when building a multi-variate cox model since cox model can only be built on limited variates. Default is -Inf (not applied).}

\item{filter.max.features}{maximum important features selected by filters. Useful when building a multi-variate cox model since cox model can only be built on limited variates. Default is Inf (not applied).}

\item{filter.result.report}{if TRUE generate filter reports for filter results, only vaild when `inbag.filter` set to FALSE (i.e. only valid in out-bag filters mode).}

\item{filter.report.all.variates}{if TRUE report all variates in the filter report, only valid when `filter.result.report` set to TRUE.}

\item{post.regression}{build a regression model based on the variates selected by LASSOBag process. Default is FALSE.}

\item{post.LASSO}{build a LASSO regression model based on the variates selected by LASSOBag process, only vaild when `post.regression` set to TRUE.}

\item{pvalue.cutoff}{determine the cut-off p-value for what variates were selected by LASSOBag, only vaild when `post.regression` is TRUE and `bagFreq.sigMethod` set to "Parametrical Statistical Test" or "Permutation Test".}

\item{used.elbow.point}{determine which elbow point to use if multiple elbow points were detectedfor what variates were selected by LASSOBag. Supported methods are "first", "middle" and "last". Default is "middle", use the middle one among all detected elbow points. Only vaild when `post.regression` is TRUE and `bagFreq.sigMethod` set to "Curve Elbow Point Detection".}
}
\value{
A list with (1) the result dataframe contains "variate" with selected frequency >1 and their "Frequency", the "P.value" and the adjusted p value "P.adjust" of each variate (if set `bagFreq.sigMethod` = "PST" or "PERT"), or the elbow point indicators "elbow.point", while elbow point(s) will be marked with "*" (if set `bagFreq.sigMethod` = "CEP"); (2) other utility results, including permutation results, the regression model built on LASSOBag results.
}
\description{
Lasso.bag is an one-step function that can be easily utilized for selecting important variates/markers from multiple models inherited from R package \emph{glmnet}. Several methods (Parametrical Statistical Test, Curve Elbow Point Detection and Permutation Test)  are provided for the cut-off point decision of the importance measure (i.e. observed selected frequency) of variates.
}
\examples{
\dontrun{
data(simulated_example)

# binomial
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$label,bootN=100,a.family="binomial",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$label,bootN=100,a.family="binomial",bagFreq.sigMethod="CEP")

# gaussian
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$y,bootN=100,a.family="gaussian",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$y,bootN=100,a.family="gaussian",bagFreq.sigMethod="CEP")

# cox
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$surv,bootN=100,a.family="cox",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$surv,bootN=100,a.family="cox",bagFreq.sigMethod="CEP")

# multinomial
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$multi.label,bootN=100,a.family="multinomial",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$multi.label,bootN=100,a.family="multinomial",bagFreq.sigMethod="CEP")

# mgaussian
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$multi.y,bootN=100,a.family="mgaussian",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$multi.y,bootN=100,a.family="mgaussian",bagFreq.sigMethod="CEP")

# poisson
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$pois,bootN=100,a.family="poisson",bagFreq.sigMethod="PST")
res<-Lasso.bag(mat=test.df$x,out.mat=test.df$pois,bootN=100,a.family="poisson",bagFreq.sigMethod="CEP")
}
}
\seealso{
\code{\link[glmnet]{glmnet}} and \code{\link[glmnet]{cv.glmnet}} in R package \emph{glmnet}.
}
