% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoBag.R
\docType{package}
\name{LassoBag}
\alias{LassoBag}
\title{LassoBag: a lasso based variable selecting CART framework}
\usage{
LassoBag(
  mat,
  out.mat,
  bootN = 1000,
  imputeN = 1000,
  boot.rep = TRUE,
  a.family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  parallel = FALSE
)
}
\arguments{
\item{mat}{independent variables, a data.frame that columns contain features while each row represents a sample.}

\item{out.mat}{dependent variables, which contains one column or two columns.}

\item{bootN}{the size of resampled subset.}

\item{imputeN}{the initial permutation times.}

\item{boot.rep}{whether sampling with return or not.}

\item{a.family}{which kind of regression method to use, it should match the type of out.mat}
}
\value{

}
\description{
LassoBag: a lasso based variable selecting CART framework

LassoBag
}
